+++ conf/urls.py
@@ -42,6 +42,11 @@ testapi_urls = [
         name='submissions_by_email',
     ),
     url(
+        r'^test-senders/$',
+        testapi.views.SendersTestAPIView.as_view(),
+        name='delete_test_senders',
+    ),
+    url(
         r'^test-submissions/$',
         testapi.views.SubmissionsTestAPIView.as_view(),
         name='delete_test_submissions',