+++ testapi/views.py
@@ -1,11 +1,11 @@
 from django.http import Http404
 from django.conf import settings
 from django.shortcuts import get_list_or_404
-from rest_framework.generics import RetrieveAPIView, DestroyAPIView
+from rest_framework.generics import RetrieveAPIView, DestroyAPIView, GenericAPIView
 from rest_framework.response import Response
 
 from client.authentication import ClientSenderIdAuthentication
-from submission.models import Submission
+from submission.models import Submission, Sender
 from submission.constants import (
     ACTION_NAME_EMAIL,
     ACTION_NAME_GOV_NOTIFY_EMAIL,
@@ -14,15 +14,18 @@ from submission.constants import (
 )
 
 
-class SubmissionsTestAPIView(DestroyAPIView, RetrieveAPIView):
-    queryset = Submission.objects.all()
+class TestAPIView(GenericAPIView):
     authentication_classes = [ClientSenderIdAuthentication]
-    http_method_names = ('delete', 'get',)
 
     def dispatch(self, *args, **kwargs):
         if not settings.FEATURE_TEST_API_ENABLED:
             raise Http404
         return super().dispatch(*args, **kwargs)
+
+
+class SubmissionsTestAPIView(TestAPIView, DestroyAPIView, RetrieveAPIView):
+    queryset = Submission.objects.all()
+    http_method_names = ('delete', 'get',)
 
     @staticmethod
     def data_and_meta(submission: Submission):
@@ -86,5 +89,21 @@ class SubmissionsTestAPIView(DestroyAPIView, RetrieveAPIView):
                     )
                     for gov_notify_action in test_gov_notify_actions:
                         gov_notify_action.delete()
+
+        return Response(status=204)
+
+
+class SendersTestAPIView(TestAPIView, DestroyAPIView):
+    queryset = Sender.objects.all()
+    http_method_names = 'delete'
+
+    def delete(self, request, **kwargs):
+        test_email_pattern = r'^test\+(.*)@directory\.uktrade\.io'
+        test_senders = get_list_or_404(
+            Sender,
+            email_address__regex=test_email_pattern,
+        )
+        for sender in test_senders:
+            sender.delete()
 
         return Response(status=204)